Upgrade to .NET 9 (closes #15, closes #11)
