--- iCon/packages.config
@@ -1,21 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="BouncyCastle.Cryptography" version="2.5.1" targetFramework="net48" />
-  <package id="HelixToolkit" version="2.27.0" targetFramework="net48" />
-  <package id="HelixToolkit.Wpf" version="2.27.0" targetFramework="net48" />
-  <package id="Microsoft.Bcl.AsyncInterfaces" version="9.0.4" targetFramework="net48" />
-  <package id="Microsoft.Extensions.DependencyInjection.Abstractions" version="9.0.4" targetFramework="net48" />
-  <package id="Microsoft.Extensions.Logging.Abstractions" version="9.0.4" targetFramework="net48" />
-  <package id="Microsoft.Xaml.Behaviors.Wpf" version="1.1.31" targetFramework="net48" />
-  <package id="PixiEditor.ColorPicker" version="3.4.2" targetFramework="net48" />
-  <package id="PixiEditor.ColorPicker.Models" version="1.0.7" targetFramework="net48" />
-  <package id="SSH.NET" version="2025.0.0" targetFramework="net48" />
-  <package id="System.Buffers" version="4.6.1" targetFramework="net48" />
-  <package id="System.Diagnostics.DiagnosticSource" version="9.0.4" targetFramework="net48" />
-  <package id="System.Formats.Asn1" version="9.0.4" targetFramework="net48" />
-  <package id="System.Memory" version="4.6.3" targetFramework="net48" />
-  <package id="System.Numerics.Vectors" version="4.6.1" targetFramework="net48" />
-  <package id="System.Runtime.CompilerServices.Unsafe" version="6.1.2" targetFramework="net48" />
-  <package id="System.Threading.Tasks.Extensions" version="4.6.3" targetFramework="net48" />
-  <package id="System.ValueTuple" version="4.6.1" targetFramework="net48" />
-</packages>
